Revert "Exercise: Data Types and Variables - completed"
This reverts commit ee73aa50b370650e96aca7a02fc1844cf5a30453.

+++ ProgrammingFundamentals/LabDataTypesAndVariables/src/CenturiesToMinutes.java
@@ -5,7 +5,7 @@ public class CenturiesToMinutes {
     public static void main(String[] args) {
         Scanner scan = new Scanner(System.in);
         int input = Integer.parseInt(scan.nextLine());
-        double years=input*100;
+      long years=input*100;
       double days=years*365.2422;
       double hours=days*24;
       double minutes=hours*60;
